refactor: split session statement handlers
